windborn_horizontal_source_left finally working as intended, creating a uniform density gradent from the top to bottom and unform snow acumulation even with lateral flux on flat ground!
step_snow_source added to my_helper to model a snow colum just out of bounds with no horozontal advection
that column is then steped along with the main body of the sim and used to caluclate windborn horozontal sorce.
This method will work as long as vertical advection is uniform.